+++ app/serializers/dnd2024/homebrew/race_serializer.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Dnd2024
+  module Homebrew
+    class RaceSerializer < ApplicationSerializer
+      attributes :id, :name, :data
+    end
+  end
+end